client: folder large-file (>4 GiB) — rc30, capability-gated folder upload guard

+++ HeidrunAppTests/ConnectionHandleTests.swift
@@ -209,7 +209,7 @@ private final class ConnectionHandleFakeClient: HotlineClient, @unchecked Sendab
     func startUpload(at path: RemotePath, name: String, size: UInt64, resume: Bool) async throws -> TransferHandle {
         TransferHandle(transferID: 0, totalSize: 0)
     }
-    func startFolderUpload(at path: RemotePath, name: String, size: UInt32, itemCount: UInt16, resume: Bool) async throws -> TransferHandle {
+    func startFolderUpload(at path: RemotePath, name: String, size: UInt64, itemCount: UInt16, resume: Bool) async throws -> TransferHandle {
         TransferHandle(transferID: 0, totalSize: 0)
     }
     func cancelTransfer(_ handle: TransferHandle) async throws {}
