+++ HeidrunAppTests/HostStateTestSupport.swift
@@ -110,7 +110,7 @@ final class FakeHotlineClient: HotlineClient, @unchecked Sendable {
     func startUpload(at path: RemotePath, name: String, size: UInt64, resume: Bool) async throws -> TransferHandle {
         TransferHandle(transferID: 0, totalSize: 0)
     }
-    func startFolderUpload(at path: RemotePath, name: String, size: UInt32, itemCount: UInt16, resume: Bool) async throws -> TransferHandle {
+    func startFolderUpload(at path: RemotePath, name: String, size: UInt64, itemCount: UInt16, resume: Bool) async throws -> TransferHandle {
         TransferHandle(transferID: 0, totalSize: 0)
     }
     func cancelTransfer(_ handle: TransferHandle) async throws {}
