+++ Packages/HeidrunModules/Tests/HeidrunChatTests/UserListViewModelTests.swift
@@ -317,7 +317,7 @@ private final class FakeUserListClient: HotlineClient, @unchecked Sendable {
     func startUpload(at path: RemotePath, name: String, size: UInt32, resume: Bool) async throws -> TransferHandle {
         TransferHandle(transferID: 0, totalSize: 0)
     }
-    func startFolderUpload(at path: RemotePath, name: String, size: UInt32, itemCount: UInt16, resume: Bool) async throws -> TransferHandle {
+    func startFolderUpload(at path: RemotePath, name: String, size: UInt64, itemCount: UInt16, resume: Bool) async throws -> TransferHandle {
         TransferHandle(transferID: 0, totalSize: 0)
     }
     func cancelTransfer(_ handle: TransferHandle) async throws {}
